AgentPart/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-ollama
langgraph
langchain-community
supabase
pytesseract
Pillow
fpdf2
playwright
python-dotenv
ollama
numpy
requests
python-multipart
